+++ decidim-core/spec/models/decidim/abilities/participatory_process_admin_ability_spec.rb
@@ -3,7 +3,7 @@
 require "spec_helper"
 
 describe Decidim::Abilities::ParticipatoryProcessAdminAbility do
-  subject { described_class.new(user, current_participatory_process: user_process) }
+  subject { described_class.new(user, current_participatory_space: user_process) }
 
   let!(:user) { create(:user, :process_admin, participatory_process: user_process) }