+++ decidim-participatory_processes/app/controllers/decidim/participatory_processes/admin/participatory_processes_controller.rb
@@ -6,6 +6,9 @@ module Decidim
       # Controller that allows managing participatory processes.
       #
       class ParticipatoryProcessesController < Decidim::Admin::ApplicationController
+        include Decidim::Admin::ParticipatorySpaceAdminContext
+        participatory_space_admin_layout only: [:edit]
+
         helper ProcessGroupsForSelectHelper
 
         helper_method :current_participatory_process
@@ -80,6 +83,10 @@ module Decidim
 
         private
 
+        def current_participatory_space
+          current_participatory_process
+        end
+
         def current_participatory_process
           @current_participatory_process ||= collection.where(slug: params[:slug]).or(
             collection.where(id: params[:slug])
@@ -91,7 +98,7 @@ module Decidim
         end
 
         def ability_context
-          super.merge(current_participatory_process: current_participatory_process)
+          super.merge(current_participatory_space: current_participatory_process)
         end
 
         def participatory_process_params